Implement screenshot recording functionality with customizable key settings and mouse click detection
